+++ src/main/java/com/telepathicgrunt/ultraamplifieddimension/blocks/GlowdirtBlock.java
@@ -1,7 +1,12 @@
 package com.telepathicgrunt.ultraamplifieddimension.blocks;
 
 import com.telepathicgrunt.ultraamplifieddimension.modInit.UADBlocks;
-import net.minecraft.block.*;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.Blocks;
+import net.minecraft.block.SnowBlock;
+import net.minecraft.block.SnowyDirtBlock;
+import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.material.MaterialColor;
 import net.minecraft.tags.FluidTags;