+++ src/main/java/com/telepathicgrunt/ultraamplifieddimension/world/features/TreeSwampHorned.java
@@ -24,7 +24,11 @@ import net.minecraft.world.gen.feature.BaseTreeFeatureConfig;
 import net.minecraft.world.gen.feature.Feature;
 import net.minecraft.world.gen.feature.template.Template;
 
-import java.util.*;
+import java.util.Comparator;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Random;
+import java.util.Set;
 
 
 public class TreeSwampHorned extends Feature<BaseTreeFeatureConfig> {